requirements.txt
streamlit
pandas
requests
scikit-learn
python-dotenv 